chatbot-flask/requirements.txt
# ── Serveur Web ─────────────────────────────────────────────
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.0

# ── HTTP / Réseau ────────────────────────────────────────────
requests==2.32.3

# ── Base de données ──────────────────────────────────────────
pymongo==4.10.1

# ── LLM Groq ─────────────────────────────────────────────────
langchain-groq==0.2.4
langchain==0.3.25
langchain-core==0.3.62
langchain-community==0.3.24

# ── Orchestration agents ─────────────────────────────────────
langgraph==0.3.34

# ── Base vectorielle (RAG) ───────────────────────────────────
chromadb==0.6.3
langchain-huggingface==0.1.2
sentence-transformers==3.4.1

# ── Utilitaires LangChain ────────────────────────────────────
tiktoken==0.9.0
numpy>=2.1.0